rename cyclops.datasets to cyclops.data
