+++ cyclops/data/__init__.py
@@ -1,0 +1,4 @@
+"""Cyclops datasets package."""
+from cyclops.data.features import MedicalImage
+from cyclops.data.packaged_loading_scripts import medical_imagefolder
+from cyclops.data.slicer import SliceSpec
