+++ cyclops/data/features/__init__.py
@@ -1,0 +1,3 @@
+"""Medical Image Feature."""
+__all__ = ["MedicalImage"]
+from cyclops.data.features.medical_image import MedicalImage
