+++ cyclops/data/features/medical_image.py
