+++ cyclops/data/packaged_loading_scripts/__init__.py
@@ -5,12 +5,12 @@ import inspect
 
 from datasets.packaged_modules import _PACKAGED_DATASETS_MODULES, _hash_python_lines
 
-from cyclops.datasets.packaged_loading_scripts.medical_imagefolder import (
+from cyclops.data.packaged_loading_scripts.medical_imagefolder import (
     medical_imagefolder,
 )
 
 # add the packaged loading scripts to huggingface datasets' registry
-# NOTE: cyclops.datasets must be imported before this change takes effect
+# NOTE: cyclops.data must be imported before this change takes effect
 _PACKAGED_DATASETS_MODULES.update(
     {
         "medicalimagefolder": (
