+++ cyclops/data/packaged_loading_scripts/medical_imagefolder/__init__.py
