--- cyclops/datasets/__init__.py
@@ -1,4 +1,0 @@
-"""Cyclops datasets package."""
-from cyclops.datasets.features import MedicalImage
-from cyclops.datasets.packaged_loading_scripts import medical_imagefolder
-from cyclops.datasets.slicer import SliceSpec
