--- cyclops/datasets/features/__init__.py
@@ -1,3 +1,0 @@
-"""Medical Image Feature."""
-__all__ = ["MedicalImage"]
-from cyclops.datasets.features.medical_image import MedicalImage
